fix: prevent duplicate loads
Vue 3 removed `this.$on` which allowed components to listen to their own
emitted events. In a much appreciated effort by another developer to
migrate this library to Vue 3, `this.$on` was replaced by a global event
bus. The problem with this is that each time `$state.loaded()` was
called on any infinite scroll, it would trigger on all others. In
particular, if you navigated back and forth from a page that was loading
slowly, this would be a problem. Consider this scenario:

You go to page with infinite scroll. Before the first page of results
get back, you navigate away. Then you navigate back and another request
is sent. When the first request returns, `$state.loaded()` is called and
a third request is sent. Now the infinite scroll is in a weird state
because it has two outstanding requests. They can come back in the wrong
order.

In this commit, the code is changed to not do emitting and listening.
Since all of the activity happens inside the component and no other
listeners are expected, we can just run the code that used to be in the